requirements.txt
streamlit
streamlit-extras
streamlit-aggrid
plotly
gradio-client
torch
torchvision
diffusers
imageio
pillow
requests